+++ Options/Defaults.lua
@@ -15,11 +15,12 @@ addon.Options.Defaults = {
         RememberSearch = false,
         RememberSearchBetweenVendors = false,
         TokenBanner = {
-            MoneyLabel = addon.L["Icon"],
-            MoneyAbbreviate = addon.L["None"],
-            ThousandsSeparator = addon.L["Space"],
+            MoneyLabel = "Icon",
+            MoneyAbbreviate = "None",
+            ThousandsSeparator = "Space",
             MoneyGoldOnly = false,
             MoneyColored = true,
+	        CurrencyAbbreviate = "None",
         }
     }
 };